Big (Logical) Update
We realised that if we invoke System.out.printl() they don't work, but we have to use the loggers to print out some debug messages and that they end up in the flink log files (and who knew that), so the run is deprecated, yes, but it is legacy and it works, so we use that.
Now, we can focus on the actual query implementations.

Co-Authored-By: chiaraiurato <92391371+chiaraiurato@users.noreply.github.com>